% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_from_plink_files.R
\name{create_data_from_plink_files}
\alias{create_data_from_plink_files}
\title{Create data.table with CHR, SNP, POS, AF, POP columns}
\usage{
create_data_from_plink_files(files_path,
  pop_name = tail(unlist(strsplit(files_path, "/")), n = 1))
}
\arguments{
\item{files_path}{string with full path of .bim and .frq files}

\item{pop_name}{string with the name of the population. DEFAULT is the
prefix of .bim and .frq files}
}
\value{
data.table with the information of .bim and .frq files
}
\description{
\code{create_data_from_plink_files()} reads .bim and .frq plink files
and merge then into one data.table
}
