% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_one_allele_per_line.R
\name{make_one_allele_perline}
\alias{make_one_allele_perline}
\title{Transform data from one SNP per line to one variant per line}
\usage{
make_one_allele_perline(freq_data)
}
\arguments{
\item{freq_data}{data.table. Data with one SNP per line and MAF information.}
}
\value{
data.table with one alelle per line with its frequency.
}
\description{
Transform data from one SNP per line to one variant per line
}
