% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_test.R
\name{moran_test}
\alias{moran_test}
\title{Moran I statitics}
\usage{
moran_test(data_in, col_name, wd_width = 2e+05, alternative = "greater")
}
\arguments{
\item{freq_list}{list of data.table. Each element of the list is a
data.table object with the allele frequency of the variants of one
population.}

\item{pheno_data}{data.table. Table with POP, mu, ssqr, NPHENO, columns
where POP has the population names, mu has the phenotypes means, ssqr has
the phenotypes standard deviations and NPHENO are the number of samples from
where the phenotypes where measured}
}
\description{
\code{popphen_conditional_mutual_information()} Calculate genetic and population mutual
information given the allele frequecies and prior probability of an idividual
to belong to one population
}
