% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgen_mutual_information.R
\name{popgen_mutual_information}
\alias{popgen_mutual_information}
\title{Calculate genetic and population mutual information.}
\usage{
popgen_mutual_information(freq_list, prior_key = "uniform")
}
\arguments{
\item{freq_list}{list of data.table. Each element of the list is a
data.table object with the allele frequency of the variants of one
population.}

\item{prior_key}{string. Gives the which prior information should be used to
calculate the mutual information (genetic or uniform). The uniform prior
probaility for one population is equal to one divided by the number of
populations. The genetic prior probability is given by the number of samples
of that population divided by the total number of samples of all
populations. Default (prior_key = 'uniform').}
}
\description{
\code{popgen_mutual_information()} Calculate genetic and population mutual
information given the allele frequecies and prior probability of an idividual
to belong to one population
}
