% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popphen_conditional_mutual_information.R
\name{popphen_conditional_mutual_information}
\alias{popphen_conditional_mutual_information}
\title{Calculate phenotipic and population mutual information conditional to the
genetic data.}
\usage{
popphen_conditional_mutual_information(freq_list, pheno_data,
  prior_key = "uniform")
}
\arguments{
\item{freq_list}{list of data.table. Each element of the list is a
data.table object with the allele frequency of the variants of one
population.}

\item{pheno_data}{data.table. Table with POP, mu, ssqr, NPHENO, columns
where POP has the population names, mu has the phenotypes means, ssqr has
the phenotypes standard deviations and NPHENO are the number of samples from
where the phenotypes where measured}

\item{prior_key}{string. Gives the which prior information should be used to
calculate the mutual information (phenotipic, genetic or uniform). The
uniform prior probaility for one population is equal to one divided by the
number of The phenotipic/genetic prior probability is given by the number of
phenotipic/genetic samples of that population divided by the total number of
samples of all populations. Default (prior_key = 'uniform').}
}
\description{
\code{popphen_conditional_mutual_information()} Calculate genetic and population mutual
information given the allele frequecies and prior probability of an idividual
to belong to one population
}
